refactor: 사이드바 섹션별로 파일 분리
